refactor(nosskey-sdk): extract shared secp256k1 helpers
liftXOnly and the shared-X derivation were duplicated between nip44.ts
and nip04.ts. Move them into secp-utils.ts as liftEvenXOnly and
ecdhSharedX so future NIPs can reuse them without copy-paste drift.

Also switch to slice() everywhere for the X-coordinate cut so we never
hand a view into noble's internal buffer to a downstream KDF / cipher.
